tbf: Add gprs_rlcmac_tbf::first_control_ts method
Currently the tbf->first_ts field is accessed directly to get the
first allocated TS when an immediate assigment is generated. While
this is in principle correct (since in that state no other TBF
exists), a more generic function that always returns the current
(first) PACCH TS would hide the details.

Add a function to return the current first PACCH timeslot.

Note that concurrent TBF are not supported yet, which will not
lead to errors yet, since it is only being used after RACH.

Sponsored-by: On-Waves ehf