tbf: Delay WAIT_ASSIGN (immediate assigment) to confirmation (TODO)
Currently the state on the TBF is set to WAIT_ASSIGN when teh
immediate assigment is created in the moment when the assignment is
really transmitted to the MS (which can take some time).

This can take some time due to PCH and AGCH queueing while the TBF
might be erroneously be considered as valid because the MS has not
yet received the message.

Delay the state change, until the BTS has confirmed it.

TODO:
 - clarify when the BTS sends the confirmation. If it is sent when
   the message is passed to the DSP, it is fine. If it is sent when
   the message is enqueued, it doesn't help much.

Sponsored-by: On-Waves ehf