tbf: Start timer when sending an immediate assigment (TODO)
Currently a timer is only started, when the confirmation is received.

Start a timer, when the assignment is sent. The timer is named T3141
(see TS 44.108, 3.5.2.1.3.1) but this might not be the correct timer
name, since TS 44.018 says, that T3141 is started, when the IMM.ASS
is sent to the MS. Here it is assumed, that this is the time the
request is passed to the BTS. The values are 10s for PCH and 5s for
AGCH.

TODO:
 - check the timeouts, sthey should be above the maximum BTS queuing
   time.

Sponsored-by: On-Waves ehf