lc15: fix configure.ac variable substitution causing compile error
When "--enable-lc15bts-phy" is passed to './configure' without specifying
an explicit header file path using "--with-litecell15=", we ended up
generating an empty string as LITECELL15_INCDIR and rendered something
like "-DENABLE_DIRECT_PHY  -I -I../../git/src/osmo-bts-litecell15" as
part of the compiler invocation, where the -I with no argument will hide
the second -I, as the second one is supposed to be the optarg for the
first include.

This in turn made the "#include <lc15_l1_if.h>" fail, when using
separate source and build directories.

This patch fixes the configur script to use $includedir, rather than the
non-existant $incdir as default for LITECELL15_INCDIR

Change-Id: I483e62f8331e7867a92f8055c4d450fdd5288cb6