Copy sysmopcu.service to osmo-pcu
This way the name of systemd service file will match the name of the
binary similar to OsmoBTS. Add aliases so the user can use both old and
new names regardless of which file is installed. Once the corresponding
changes to OE recipes are applied old file can be removed.

Based on work by Pau Espin Pedrol <pespin@sysmocom.de>

Change-Id: I2ca6f6c486bd6fcf4d5b3d0a05d25dc04f020c26